+++ CHANGELOG.md
@@ -16,6 +16,7 @@ All notable changes to this project are documented in this file.
  - Added a `cache-rendering-hint` boolean property that can be applied to any element, to hint to the renderer that it should cache the element and its children
    into a cached layer. This may speed up rendering of complex sub-trees if they rarely change.
  - The `brighter` and `lighter` functions also work on values of type brush.
+ - Added a `reset` function to C++'s `Model`, Rust's `ModelNotify` and JS's `ModelPeer`
 
 ### Fixed